// v2.client/src/shared/api/generated/models/discord-cmd.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DiscordCmd {
    'id'?: string;
    'insDate'?: string;
    'updDate'?: string;
    'insBy'?: string | null;
    'updBy'?: string | null;
    'cmd': string | null;
    'response'?: string | null;
    'active'?: boolean;
}